Consumer shutdown improvements, long job support
The consumer close call's completion is now waiting for. Does seem to
have improved the stabilisation problem, but not solved it. The consumer
will now also support long processing tasks, by pausing the kafka client
and polling it to keep alive it's partition assignment. Moved over to
auto offset commit, it's sent on every poll and we don't check for
shutdown until all read tasks are completed. Note: We could reduce our
max.poll.records to bring this in tighter.