Command ids used to select shard, multi events sequenced
When putting messages onto Kinesis streams the command id, if
available, is now used. This ensures that commands and the events they
create will be on the same shard. When a command handler returns multi
events a sequence number is added to ensure that the order is
maintained.

Together these changes should ensure consumers receive events and
commands in the order they are submitting to the streams.