Remove code where invalid commands (spec/valid?) cause an exception
We would like command handlers to generate more context aware
rejection/invalid events when something in the command is
wrong. Commands must have achieved a basic level of validation by this
point to have even been matched with a command handler. Command
handlers should therefore attempt to diagnose what is invalid about a
command and respond with an appropriate event, such as a rejection
event, with better information for the user about why the rejection occurred.