Unhandled exceptions now cause the kinesis consumer to be shutdown
Previously we continued processing on unexpected exceptions, but its
clear now that this could lead to problems with our event aggregators.

This change removes catches from within the code and allows exceptions
to bubble up to the top level Kinesis processor function. This logs a
fatal level message, before triggering the worker that is wrapping it
to shutdown. This comms is handled via a dedicated core.async channel.

This should leave the rest of any service using the library able to
handle query requests and the other processor be that for events or
commands. It is expected that Kinesis, upon seeing this processor
shutdown, will allocate the current shards over to other processors,
causing them to fail as well.

Alerts have been configured for all fatal level messages, and the
message will include the offending Kinesis record data. It is expected
that some kind of manual fix will be required and then the service
will need to be restarted.